Keep lighting changes in an intelligent stack
By stacking changes on top of each other, it is possible to have multiple overlapping events active at the same time. The intelligent part comes into play when removing an event from the top or even middle of the stack. On such an event, the system starts a top-down search for the new state for each device that was effected by the old state. This way, it is impossible to have a device become stuck in a state it shouldn't be in due to multiple simultaneous events